+++ rororo/annotations.py
@@ -1,0 +1,14 @@
+"""
+==================
+rororo.annotations
+==================
+
+Internal module to keep all reusable type annotations for the project in one
+place.
+
+"""
+
+from typing import Any, Dict
+
+
+StrAnyDict = Dict[str, Any]